requirements.txt
streamlit
pandas
pandasai
matplotlib
langchain-google-genai
google-generativeai
